[DEV-9734b] create Confirm and Error components
